+++ ThirdPartyAdapters/mytarget/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 #### Next Version
 - Maps `AgeRestrictedTreatment` to myTarget's age restriction settings.
-- Updated interstitial ad implementation to use updated mediation adapter APIs.
+- Updated interstitial and native ad implementation to use updated mediation adapter APIs.
 
 #### Version 5.47.1.0
 - Added property to build the adapter with GMA Next-Gen SDK dependency.